ec2: add separate waiting for update function for C2 volumes
This fixes an issue with completing C2 volume update due to
VolumeModification not found.

C2 doesn't implement VolumeModification, so updates of C2 volumes can't
be tracked with default wait functions like `WaitVolumeUpdated` or
`WaitVolumeModificationComplete` because of differences in states model.

`detaching` is added to the volume state set because it is one of the
states in which C2 volume can be modified. It is used as a target state.