paas: change type Int -> NullableInt for parameters contained big ints
The problem with big values in tf configurations can be solved using
string as a parameter type with a conversion to int64. This approach is
suggested in the discussion:
https://github.com/hashicorp/terraform-plugin-sdk/issues/92

NullableInt was chosen because it is based on string type, can hold
int64 values, and already implements methods for string-int64
conversion.

Changing the type also allows to bypass the bug with specifying max
int64 value in the configuration.